Make changes to work and build WindowsDesktop
This is a squashed mix of changes that make the infra use the new sharedfx tooling SDK, trim out obviously unnecessary code, and build bits that seem to work.

Unfortunately, this does contain a few segments that are copied from Core-Setup. Because the directory structure of this repo is different, some files don't exist, so the contents had to move to an existing file.