Re-add inbuild signing (#3024)
* Re-add inbuild signing
This re-enables inbuild signing. Unlike the previous iteration of inbuild signing we now can use the same approach for dealing with containers as we do with post-build signing (sign wixpacks, repack them, iteratively). Essentially a way to think of this is that we are doing post-build signing, but it happens to be in-build.

To do so, a target is added in the prepare-artifacts project that calls the arcade sign project (windowsdesktop doesn't really use the standard arcade process).

* Fixup some comments

* Explicitly set UseDotNetCertificate