Replace arch-specific package properties with arch neutral ones (#5081)
This avoids need for a workaround in VMR builds and reduces friction for stable builds. When building non-win-x64 verticals, we end up having to set the x64 runtime pack and x64 VS redist pack properties (stable and non-stable properties) based on other values in the VMR orchestrator. Instead, use MicrosoftNEtCoreAppRef as the stable version property, and MicrosoftNETCOrePlatforms as the non-stable one.